+++ src/app/domains/categories/domains.transaction-categories.types.ts
@@ -1,5 +1,5 @@
-import { TServerEntityId } from "src/app/common/http/common.http.types";
-import { WalletTransactionType } from "../transactions/domains.transactions.constants";
+import { TServerEntityId } from 'src/app/common/http/common.http.types';
+import { WalletTransactionType } from '../transactions/domains.transactions.constants';
 
 export type CategorySelectionValue = WalletTransactionType | '';
 
@@ -9,14 +9,3 @@ export interface ITransactionCategoryApiResponse {
   type: WalletTransactionType;
 }
 
-export class TransactionCategory {
-  public readonly id: TServerEntityId | null;
-  public readonly name: string;
-  public readonly type: WalletTransactionType;
-
-  constructor(id: TServerEntityId | null, name: string, type: WalletTransactionType) {
-    this.id = id;
-    this.name = name;
-    this.type = type;
-  }
-}
