+++ src/app/pages/categories/management/pages-transaction-categories.service.ts
@@ -1,21 +1,21 @@
 import { Injectable } from '@angular/core';
 import { map, Observable } from 'rxjs';
 import { DomainsTransactionCategoriesGateway } from 'src/app/domains/categories/domains-transaction-categories-gateway';
-import { TransactionCategory } from 'src/app/domains/categories/domains.transaction-categories.types';
+import { TransactionCategory } from '../transaction-category.model';
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
 export class PagesTransactionCategoriesService {
-
-  constructor(
+  public constructor(
     private readonly domainsTransactionCategoriesGateway: DomainsTransactionCategoriesGateway,
   ) { }
 
   public getTransactionCategories(): Observable<TransactionCategory[]> {
     return this.domainsTransactionCategoriesGateway.getTransactionCategories().pipe(
-      map(categories => categories.map(c => new TransactionCategory(c.id, c.name, c.type )))
+      map(categories => categories.map(
+        category => new TransactionCategory(category.id, category.name, category.type ),
+      )),
     );
   }
-
 }
